+++ doc/CHANGELOG.md
@@ -1,6 +1,9 @@
 
 ## Latest changes
 
+* Unstructured constellation polygon creation: Fix vehicle curvature calculation after response time, make DebugDrawing accessible from python
+
+
 ## Release 4.2.0
 
 #### :ghost: Maintenance